+++ src/features/TextToImage/config/types/Config.ts
@@ -1,4 +1,4 @@
-import type * as WebAPI from '@/library/WebAPI';
+import type * as WebAPI from '@/library/WebAPI/exports';
 
 /** The user-provided settings for the text-to-image feature */
 class TextToImage_Config {
